# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "My Library API",
    "version": "0.2.0",
    "description": "Books, categories and ratings of the home library. Errors use RFC 9457 Problem Details."
  },
  "servers": [
    {
      "url": "https://library.bmcpthilakawansha.workers.dev",
      "description": "production"
    },
    {
      "url": "http://localhost:8787",
      "description": "local wrangler dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-session",
        "description": "Set by /auth/login → Asgardeo → /auth/callback"
      }
    },
    "schemas": {
...
      "BookPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Book"
            }
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "pages"
        ]
      },
...
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "example": "Bad Request"
          },
          "status": {
            "type": "integer",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "pageSize: Too big: expected number to be <=100"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "field",
                "message"
              ]
            }
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "bookCount": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "bookCount"
        ]
      },
      "Language": {
        "type": "object",
        "properties": {
          "language": {
            "type": [
              "string",
              "null"
            ]
          },
          "bookCount": {
            "type": "integer"
          }
        },
        "required": [
          "language",
          "bookCount"
        ]
      },
...
      "LoanRow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "bookId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "titleSinglish": {
            "type": [
              "string",
              "null"
            ]
          },
          "coverUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "borrowedAt": {
            "type": "string",
            "format": "date-time"
          },
          "returnedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "lentByName": {
            "type": [
              "string",
              "null"
            ]
          },
          "days": {
            "type": "integer"
          },
          "overdue": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "bookId",
          "title",
          "titleSinglish",
          "coverUrl",
          "userId",
          "name",
          "email",
          "borrowedAt",
          "returnedAt",
          "lentByName",
          "days",
          "overdue"
        ]
      },
...
      "LendBody": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "bookId",
          "userId"
        ]
      },
...
      "ShelfCounts": {
        "type": "object",
        "properties": {
          "to_read": {
            "type": "integer"
          },
          "reading": {
            "type": "integer"
          },
          "read": {
            "type": "integer"
          },
          "rated": {
            "type": "integer"
          }
        },
        "required": [
          "to_read",
          "reading",
          "read",
          "rated"
        ]
      },
      "Recommendation": {
        "type": "object",
        "properties": {
          "book": {
            "$ref": "#/components/schemas/Book"
          },
          "score": {
            "type": "number",
            "description": "0–1, higher = better match"
          },
          "reason": {
            "type": "string",
            "example": "Because you liked other books by Martin Wickramasinghe"
          }
        },
        "required": [
          "book",
          "score",
          "reason"
        ]
      }
    },
...
  },
  "paths": {
...
    "/api/books": {
      "get": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "Books"
        ],
        "operationId": "listBooks",
        "summary": "List books (search, filter, sort, page)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "maxLength": 100,
              "description": "Search title, author, translator… (Sinhala, English or Singlish)",
              "example": "madol"
            },
            "required": false,
            "description": "Search title, author, translator… (Sinhala, English or Singlish)",
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 40,
              "example": "Sinhala"
            },
            "required": false,
            "name": "language",
            "in": "query"
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              ],
              "description": "Repeat for several: ?category=Novel&category=War (a book needs at least one)"
            },
            "required": false,
            "description": "Repeat for several: ?category=Novel&category=War (a book needs at least one)",
            "name": "category",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "true = only translated books"
            },
            "required": false,
            "description": "true = only translated books",
            "name": "translations",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "true = only books with a free copy"
            },
            "required": false,
            "description": "true = only books with a free copy",
            "name": "available",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "to_read",
                "reading",
                "read",
                "rated",
                "borrowed"
              ],
              "description": "Only books on MY shelf (or that I rated)"
            },
            "required": false,
            "description": "Only books on MY shelf (or that I rated)",
            "name": "shelf",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "true = leave out books I have read"
            },
            "required": false,
            "description": "true = leave out books I have read",
            "name": "hideRead",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "title",
                "newest",
                "rating"
              ],
              "default": "title"
            },
            "required": false,
            "name": "sort",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 24
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "One page of books",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookPage"
                }
              }
            }
          },
          "400": {
            "description": "The query string is not valid",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Account not active yet",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/categories": {
      "get": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "Meta"
        ],
        "operationId": "listCategories",
        "summary": "All categories with how many books each has",
        "responses": {
          "200": {
            "description": "Categories A–Z",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/languages": {
      "get": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "Meta"
        ],
        "operationId": "listLanguages",
        "summary": "Book languages with counts",
        "responses": {
          "200": {
            "description": "Languages A–Z",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Language"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/loans": {
      "get": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "Admin: lending"
        ],
        "operationId": "adminListLoans",
        "summary": "Loans (open first, longest out first)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "overdue",
                "returned"
              ],
              "default": "open"
            },
            "required": false,
            "name": "status",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Loans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LoanRow"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "Admin: lending"
        ],
        "operationId": "adminLendBook",
        "summary": "Lend a copy of a book to a reader",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LendBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanRow"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such book or reader",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "No free copy, reader not active, or they already have it",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me/loans": {
      "get": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "My reading"
        ],
        "operationId": "myLoans",
        "summary": "Books I have now, and my last 20 returned",
        "responses": {
          "200": {
            "description": "My loans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "open": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LoanRow"
                      }
                    },
                    "history": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LoanRow"
                      }
                    },
                    "loanDays": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "open",
                    "history",
                    "loanDays"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/books": {
      "post": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "Admin: books"
        ],
        "operationId": "adminCreateBook",
        "summary": "Add a book",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Same ISBN already in the library",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/covers": {
      "put": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "Admin: AI & covers"
        ],
        "operationId": "adminUploadCover",
        "summary": "Store a (polished) cover image in R2",
        "requestBody": {
          "required": true,
          "content": {
            "image/jpeg": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Stored",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string"
                    },
                    "bytes": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "url",
                    "bytes"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Image larger than 2 MB",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "415": {
            "description": "Not a JPEG, PNG or WebP image",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "AI free quota used up",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "AI error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "AI busy",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me/shelves": {
      "get": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "My reading"
        ],
        "operationId": "myShelves",
        "summary": "How many books are on each of my shelves",
        "responses": {
          "200": {
            "description": "Counts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShelfCounts"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/recommendations": {
      "get": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "My reading"
        ],
        "operationId": "myRecommendations",
        "summary": "Books I have not touched yet, best match first, each with a reason",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 24,
              "default": 8
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Recommendations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Recommendation"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    }
  },
...
}